chore: Move and rename some files
